+++ lib-ws28xx/src/h3/ws28xxmulti.cpp
@@ -2,7 +2,7 @@
  * @file ws28xxmulti.cpp
  *
  */
-/* Copyright (C) 2019-2023 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2019-2024 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -56,7 +56,7 @@ WS28xxMulti::WS28xxMulti(PixelConfiguration& pixelConfiguration): m_PixelConfigu
 	assert(s_pThis == nullptr);
 	s_pThis = this;
 
-	uint16_t nLedsPerPixel;
+	uint32_t nLedsPerPixel;
 	m_PixelConfiguration.Validate(nLedsPerPixel);
 
 	const auto nCount = m_PixelConfiguration.GetCount();
