+++ lib-ws28xxdmx/src/dmxmulti/ws28xxdmxmulti.cpp
@@ -108,7 +108,7 @@ void WS28xxDmxMulti::SetData(const uint32_t nPortIndex, const uint8_t* pData, ui
 
 	const auto nGroups = m_pixelDmxConfiguration.GetGroups();
 	const auto beginIndex = m_PortInfo.nBeginIndexPort[nSwitch];
-	const auto endIndex = std::min(nGroups, static_cast<uint16_t>(beginIndex + (nLength / m_nChannelsPerPixel)));
+	const auto endIndex = std::min(nGroups, (beginIndex + (nLength / m_nChannelsPerPixel)));
 
 	uint32_t d = 0;
